@@ -5,7 +5,7 @@ import math
 import shutil
 import tempfile
 from pathlib import Path
-from typing import Literal, Optional
+from typing import Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -27,7 +27,7 @@ def generate_reports(results_path: Path, output_dir: Path):
     benchmark_df = pd.read_csv(results_path)
 
     def format_run_details(input: dict) -> str:
-        if input is None or input == np.nan:
+        if input is None or input == np.nan:  # noqa: PLW0177
             return ""
         if isinstance(input, str):
             input = ast.literal_eval(input)
@@ -109,8 +109,8 @@ def save_chart(
     df: pd.DataFrame,
     title: str,
     output_path: Path,
-    yscale: Optional[Literal["linear", "log", "symlog", "logit"]] = None,
-    y_value_formatter: Optional[str] = None,
+    yscale: Literal["linear", "log", "symlog", "logit"] | None = None,
+    y_value_formatter: str | None = None,
     print_labels_on_points: bool = False,
     size: tuple[float, float] = (8, 4),
     plot_kind: Literal[
@@ -126,8 +126,8 @@ def save_chart(
         "scatter",
         "hexbin",
     ] = "line",
-    gridlines: Optional[Literal["both", "x", "y"]] = None,
-    linestyle: Optional[str] = None,
+    gridlines: Literal["both", "x", "y"] | None = None,
+    linestyle: str | None = None,
 ):
     """Render and save a chart.
 
@@ -246,7 +246,7 @@ def save_chart(
     if not output_path.exists():
         fig.savefig(str(output_path))
     else:
-        # If it exists already, only save it if it has changed
+        # If it already exists, only save it if it has changed
         with tempfile.TemporaryDirectory() as tmp_dir:
             tmp_output_path = Path(tmp_dir) / output_path.name
             fig.savefig(tmp_output_path)
