@@ -8,14 +8,16 @@ import shutil
 import sqlite3
 import time
 import warnings
+from collections.abc import Iterable
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any, Union
 
 from pygeoops import GeometryType
+from pyproj import CRS, Transformer
 
 import geofileops as gfo
 from geofileops.helpers._configoptions_helper import ConfigOptions
-from geofileops.util import _sqlite_userdefined as sqlite_userdefined
+from geofileops.util import _geopath_util, _sqlite_userdefined
 from geofileops.util._general_util import MissingRuntimeDependencyError
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -91,12 +93,117 @@ class SqliteProfile(enum.Enum):
     SPEED = 1
 
 
+def add_gpkg_ogr_contents(database: Any, layer: str | None, force_update: bool = False):
+    """Add a layer to the gpkg_ogr_contents table in a geopackage.
+
+    If the table doesn't exist yet, it will be created.
+
+    Args:
+        database (Any): the database to add the gpkg_ogr_contents to. This can be a path
+            to a file or an sqlite3.Connection object. If it is a connection, it won't
+            be closed when returning.
+        layer (str): the name of the layer to add to the gpkg_ogr_contents table.
+            If None, only the table will be created.
+        force_update (bool, optional): True to update the data if the layer already
+            exists. Defaults to False.
+
+    Raises:
+        RuntimeError: an error occured while creating the table.
+
+    Returns:
+        None
+    """
+    if isinstance(database, sqlite3.Connection):
+        # If a connection is passed, use it
+        conn = database
+    else:
+        conn = sqlite3.connect(database)
+
+    sql = None
+    try:
+        # Create gpkg_ogr_contents table if it doesn't exist yet
+        sql = """
+            CREATE TABLE IF NOT EXISTS gpkg_ogr_contents(
+                table_name TEXT NOT NULL PRIMARY KEY,
+                feature_count INTEGER DEFAULT 0
+            );
+        """
+        conn.execute(sql)
+
+        if layer is None:
+            conn.commit()
+            return
+
+        # Create triggers to keep the feature count up to date
+        sql = f"""
+            CREATE TRIGGER IF NOT EXISTS "trigger_insert_feature_count_{layer}"
+                AFTER INSERT ON "{layer}"
+            BEGIN
+                UPDATE gpkg_ogr_contents
+                    SET feature_count = feature_count + 1
+                WHERE lower(table_name) = lower('{layer}');
+            END;
+        """
+        conn.execute(sql)
+        sql = f"""
+            CREATE TRIGGER IF NOT EXISTS "trigger_delete_feature_count_{layer}"
+                AFTER DELETE ON "{layer}"
+            BEGIN
+                UPDATE gpkg_ogr_contents
+                    SET feature_count = feature_count - 1
+                WHERE lower(table_name) = lower('{layer}');
+            END;
+        """
+        conn.execute(sql)
+
+        # Check if the layer already exists in the gpkg_ogr_contents table
+        sql = f"""
+            SELECT * FROM gpkg_ogr_contents
+             WHERE lower(table_name) = '{layer.lower()}';
+        """
+        conn.row_factory = sqlite3.Row
+        cursor = conn.execute(sql)
+        contents = cursor.fetchall()
+        if len(contents) > 0:
+            # Layer already exists, check if we need to update it
+            if force_update:
+                sql = f"""
+                    UPDATE gpkg_ogr_contents
+                       SET feature_count = (SELECT COUNT(*) FROM "{layer}")
+                     WHERE lower(table_name) = '{layer.lower()}';
+                """
+                conn.execute(sql)
+            else:
+                # Layer already exists and we don't want to update it, so skip it
+                return
+        else:
+            # Layer doesn't exist yet, so add it to the gpkg_ogr_contents table
+            sql = f"""
+                INSERT INTO gpkg_ogr_contents (table_name, feature_count)
+                  VALUES ('{layer}', (SELECT COUNT(*) FROM "{layer}"));
+            """
+            conn.execute(sql)
+
+        conn.commit()
+
+    except Exception as ex:  # pragma: no cover
+        conn.rollback()
+        raise RuntimeError(f"Error executing {sql}") from ex
+    finally:
+        # If no existing connection was passed, close the connection
+        if not isinstance(database, sqlite3.Connection):
+            conn.close()
+
+
 def create_new_spatialdb(
     path: Union[str, "os.PathLike[Any]"],
-    crs_epsg: Optional[int] = None,
-    filetype: Optional[str] = None,
+    crs_epsg: int | None = None,
+    filetype: str | None = None,
 ) -> sqlite3.Connection:
     """Create a new spatialite database file.
+
+    Notes:
+        - the bounds filled out in the gpkg_contents table will be the bounds of the crs
 
     Args:
         path (PathLike): the path the create the database file.
@@ -198,7 +305,7 @@ def create_new_spatialdb(
     except ValueError:
         conn.close()
         raise
-    except Exception as ex:
+    except Exception as ex:  # pragma: no cover
         conn.close()
         raise RuntimeError(f"Error creating spatial db {path} executing {sql}") from ex
 
@@ -210,23 +317,28 @@ def get_columns(
     input_databases: dict[str, Path],
     empty_output_ok: bool = True,
     use_spatialite: bool = True,
-    output_geometrytype: Optional[GeometryType] = None,
+    output_geometrytype: GeometryType | None = None,
 ) -> dict[str, str]:
     # Init
     start = time.perf_counter()
     tmp_dir = None
 
+    def get_filetype(path: Path) -> str:
+        if _geopath_util.suffixes(path).lower() in (".gpkg", ".gpkg.zip"):
+            return "gpkg"
+        return path.suffix.lstrip(".")
+
     # Connect to/create sqlite main database
     if "main" in input_databases:
         # If an input database is main, use it as the main database
         main_db_path = input_databases["main"]
-        filetype = main_db_path.suffix.lstrip(".")
+        filetype = get_filetype(main_db_path)
         conn = sqlite3.connect(main_db_path, detect_types=sqlite3.PARSE_DECLTYPES)
         new_db = False
     else:
         # Create temp output db to be sure the output DB is writable, even though we
         # only create a temporary table.
-        filetype = next(iter(input_databases.values())).suffix.lstrip(".")
+        filetype = get_filetype(next(iter(input_databases.values())))
         conn = create_new_spatialdb(":memory:", filetype=filetype)
         new_db = True
 
@@ -352,7 +464,7 @@ def get_columns(
             else:
                 columns[columnname] = columntype
 
-    except Exception as ex:
+    except Exception as ex:  # pragma: no cover
         conn.rollback()
         raise RuntimeError(f"Error {ex} executing {sql}") from ex
     finally:
@@ -368,18 +480,121 @@ def get_columns(
     return columns
 
 
+def copy_table(
+    input_path: Union[str, "os.PathLike[Any]"],
+    output_path: Union[str, "os.PathLike[Any]"],
+    input_table: str,
+    output_table: str,
+    columns: Iterable[str] | None = None,
+    where: str | None = None,
+    preserve_fid: bool = False,
+    profile: SqliteProfile = SqliteProfile.DEFAULT,
+) -> None:
+    """Copy data from one to another table.
+
+    Notes:
+        - At the moment only appending to an existing table is supported.
+        - At the moment only copying from one sqlite file to another is supported.
+
+    Args:
+        input_path (PathLike): The path to the input SQLite database.
+        output_path (PathLike): The path to the output SQLite database.
+        input_table (str): The name of the input table.
+        output_table (str): The name of the output table.
+        columns (Iterable[str] | None, optional): The list of columns to copy. If None,
+            all columns will be copied.
+        where (str | None, optional): An optional SQL WHERE clause to filter the rows
+            to copy. Defaults to None.
+        preserve_fid (bool, optional): Whether to preserve values in the input FID
+            column. Defaults to False.
+        profile (SqliteProfile, optional): The SQLite profile to use.
+            Defaults to SqliteProfile.DEFAULT.
+    """
+    # copy_table only supports local paths... so we can just use Path
+    input_path = Path(input_path)
+    output_path = Path(output_path)
+    if not input_path.exists():
+        raise FileNotFoundError(f"Input file not found: {input_path}")
+    if not output_path.exists():
+        raise FileNotFoundError(f"Output file not found: {output_path}")
+    if input_path.resolve() == output_path.resolve():
+        raise ValueError(f"Input and output paths cannot be the same: {input_path}")
+
+    conn = sqlite3.connect(str(output_path), uri=True)
+
+    # Execute the insert statement
+    sql = None
+    try:
+        load_spatialite(conn)
+
+        if Path(output_path).suffix.lower() == ".gpkg":
+            sql = "SELECT EnableGpkgMode();"
+            conn.execute(sql)
+
+        # Attach the input database
+        sql = "ATTACH DATABASE ? AS input_db"
+        dbSpec = (str(input_path),)
+        conn.execute(sql, dbSpec)
+
+        # Set some default performance options
+        database_names = ["main", "input_db"]
+        set_performance_options(conn, profile, database_names)
+
+        # Start transaction manually needed for performance
+        sql = "BEGIN TRANSACTION;"
+        conn.execute(sql)
+
+        if columns is None:
+            # If the columns are not specified, determine them from the input table.
+            # If the input layer has fewer columns than the output, those columns will
+            # simply get the default values...
+            # If the fid should not be preserved, don't include it in the column list
+            column_filter = "" if preserve_fid else "WHERE lower(name) <> 'fid'"
+            sql = f"""
+                SELECT name
+                FROM pragma_table_info('{input_table}', 'input_db')
+                {column_filter};
+            """
+            columns = [value[0] for value in conn.execute(sql).fetchall()]
+
+        elif preserve_fid and "fid" not in [col.lower() for col in columns]:
+            # If preserve_fid is asked, the fid should be in the list of columns.
+            columns = list(columns)
+            columns.append("fid")
+
+        columns_str = ", ".join([f'"{col}"' for col in columns])
+        where_clause = f"WHERE {where}" if where else ""
+        sql = f"""
+            INSERT INTO main."{output_table}"
+                ({columns_str})
+            SELECT {columns_str}
+              FROM input_db."{input_table}"
+             {where_clause};
+        """
+        conn.execute(sql)
+
+        conn.commit()
+    except Exception as ex:  # pragma: no cover
+        conn.rollback()
+
+        raise RuntimeError(f"Error {ex} executing {sql}") from ex
+    finally:
+        conn.close()
+
+
 def create_table_as_sql(
     input_databases: dict[str, Path],
     output_path: Path,
     sql_stmt: str,
     output_layer: str,
-    output_geometrytype: Optional[GeometryType],
+    output_geometrytype: GeometryType | None,
     output_crs: int,
     append: bool = False,
     update: bool = False,
     create_spatial_index: bool = False,
+    create_ogr_contents: bool = False,
     empty_output_ok: bool = True,
-    column_datatypes: Optional[dict] = None,
+    column_datatypes: dict | None = None,
     profile: SqliteProfile = SqliteProfile.DEFAULT,
 ):
     """Execute sql statement and save the result in the output file.
@@ -396,6 +611,8 @@ def create_table_as_sql(
         update (bool, optional): True to append to an existing layer. Defaults to False.
         create_spatial_index (bool, optional): True to create a spatial index on the
             output layer. Defaults to False.
+        create_ogr_contents (bool, optional): True to create the gpkg_ogr_contents table
+            in the output file and fill it up. Defaults to False.
         empty_output_ok (bool, optional): If the sql_stmt doesn't return any rows and
             True, create an empty output file. If False, throw EmptyResultError.
             Defaults to True.
@@ -529,13 +746,30 @@ def create_table_as_sql(
         if output_suffix_lower == ".gpkg":
             data_type = "features" if "geom" in column_types else "attributes"
 
+            # Fill out the bounds of the layer using the bounds of the crs if possible
+            try:
+                crs = CRS.from_user_input(output_crs)
+                if crs is not None and crs.area_of_use is not None:
+                    transformer = Transformer.from_crs(
+                        crs.geodetic_crs, crs, always_xy=True
+                    )
+                    bounds = transformer.transform_bounds(*crs.area_of_use.bounds)
+                    min_x, min_y, max_x, max_y = (
+                        to_string_for_sql(coord) for coord in bounds
+                    )
+                else:
+                    min_x = min_y = max_x = max_y = "NULL"
+
+            except Exception:
+                min_x = min_y = max_x = max_y = "NULL"
+
             # ~ mimic behaviour of gpkgAddGeometryColumn()
             sql = f"""
                 INSERT INTO {output_databasename}.gpkg_contents (
-                    table_name, data_type, identifier, description, last_change,
-                    min_x, min_y, max_x, max_y, srs_id)
+                  table_name, data_type, identifier, description, last_change,
+                  min_x, min_y, max_x, max_y, srs_id)
                 VALUES ('{output_layer}', '{data_type}', NULL, '', DATETIME(),
-                    NULL, NULL, NULL, NULL, {to_string_for_sql(output_crs)});
+                  {min_x}, {min_y}, {max_x}, {max_y}, {to_string_for_sql(output_crs)});
             """
             conn.execute(sql)
 
@@ -571,7 +805,7 @@ def create_table_as_sql(
             columns_for_insert = [f'"{column}"' for column in column_types]
             sql = (
                 f'INSERT INTO {output_databasename}."{output_layer}" '
-                f'({", ".join(columns_for_insert)})\n{sql_stmt}'
+                f"({', '.join(columns_for_insert)})\n{sql_stmt}"
             )
             conn.execute(sql)
 
@@ -586,8 +820,12 @@ def create_table_as_sql(
                 )
             raise
 
+        # Fill out the feature count in the gpkg_ogr_contents table + create triggers
+        if create_ogr_contents and output_suffix_lower == ".gpkg":
+            add_gpkg_ogr_contents(database=conn, layer=output_layer, force_update=True)
+
         # Create spatial index if needed
-        if "geom" in column_types and create_spatial_index:
+        if create_spatial_index and "geom" in column_types:
             sql = f"SELECT UpdateLayerStatistics('{output_layer}', 'geom');"
             conn.execute(sql)
             if output_suffix_lower == ".gpkg":
@@ -621,16 +859,14 @@ def create_table_as_sql(
         conn.close()
         if output_path.exists():
             output_path.unlink()
-    except Exception as ex:
+    except Exception as ex:  # pragma: no cover
         raise RuntimeError(f"Error {ex} executing {sql}") from ex
     finally:
         if conn is not None:
             conn.close()
 
 
-def execute_sql(
-    path: Path, sql_stmt: Union[str, list[str]], use_spatialite: bool = True
-):
+def execute_sql(path: Path, sql_stmt: str | list[str], use_spatialite: bool = True):
     # Connect to database file
     conn = sqlite3.connect(path)
     sql = None
@@ -659,9 +895,82 @@ def execute_sql(
 
     except Exception as ex:
         conn.rollback()
-        raise Exception(f"Error executing {sql}") from ex
+        raise RuntimeError(f"Error executing {sql}") from ex
     finally:
         conn.close()
+
+
+def get_gpkg_contents(path: Path) -> dict[str, dict]:
+    """Get the contents of the gpkg_contents table of a geopackage.
+
+    Args:
+        path (Path): file path to the geopackage.
+
+    Returns:
+        dict[str, Any]: the contents of the geopackage.
+    """
+    conn = sqlite3.connect(path)
+    sql = None
+    try:
+        sql = "SELECT * FROM gpkg_contents;"
+        conn.row_factory = sqlite3.Row
+        cursor = conn.execute(sql)
+        contents = cursor.fetchall()
+        contents_dict = {row["table_name"]: dict(row) for row in contents}
+    except Exception as ex:  # pragma: no cover
+        raise RuntimeError(f"Error executing {sql}") from ex
+    finally:
+        conn.close()
+
+    return contents_dict
+
+
+def get_gpkg_ogr_contents(path: Path) -> dict[str, dict]:
+    """Get the contents of the gpkg_ogr_contents table of a geopackage.
+
+    Args:
+        path (Path): file path to the geopackage.
+
+    Returns:
+        dict[str, Any]: the contents of the geopackage.
+    """
+    conn = sqlite3.connect(path)
+    sql = None
+    try:
+        sql = "SELECT * FROM gpkg_ogr_contents;"
+        conn.row_factory = sqlite3.Row
+        cursor = conn.execute(sql)
+        contents = cursor.fetchall()
+        contents_dict = {row["table_name"]: dict(row) for row in contents}
+    except Exception as ex:  # pragma: no cover
+        raise RuntimeError(f"Error executing {sql}") from ex
+    finally:
+        conn.close()
+
+    return contents_dict
+
+
+def get_tables(path: Path) -> list[str]:
+    """List all tables in the database.
+
+    Args:
+        path (Path): file path to the database.
+
+    Returns:
+        list[str]: the list of all tables in the database.
+    """
+    conn = sqlite3.connect(path)
+    sql = None
+    try:
+        sql = "SELECT name FROM sqlite_master WHERE type='table';"
+        cursor = conn.execute(sql)
+        tables = [row[0] for row in cursor.fetchall()]
+    except Exception as ex:  # pragma: no cover
+        raise RuntimeError(f"Error executing {sql}") from ex
+    finally:
+        conn.close()
+
+    return tables
 
 
 def test_data_integrity(path: Path, use_spatialite: bool = True):
@@ -694,15 +1003,15 @@ def test_data_integrity(path: Path, use_spatialite: bool = True):
                     # All data was fetched from layer
                     break
 
-    except Exception as ex:
-        raise Exception(f"Error executing {sql}") from ex
+    except Exception as ex:  # pragma: no cover
+        raise RuntimeError(f"Error executing {sql}") from ex
     finally:
         conn.close()
 
 
 def set_performance_options(
     conn: sqlite3.Connection,
-    profile: Optional[SqliteProfile] = None,
+    profile: SqliteProfile | None = None,
     database_names: list[str] = [],
 ):
     try:
@@ -713,7 +1022,7 @@ def set_performance_options(
         sql = "PRAGMA temp_store=2;"
         conn.execute(sql)
         # Set soft heap limit to 1 GB (in bytes)
-        sql = f"PRAGMA soft_heap_limit={1024*1024*1024};"
+        sql = f"PRAGMA soft_heap_limit={1024 * 1024 * 1024};"
         conn.execute(sql)
 
         # Use the sqlite profile specified
@@ -735,7 +1044,7 @@ def set_performance_options(
                 sql = f"PRAGMA {databasename}.synchronous=OFF;"
                 conn.execute(sql)
 
-    except Exception as ex:
+    except Exception as ex:  # pragma: no cover
         raise RuntimeError(f"Error executing {sql}: {ex}") from ex
 
 
@@ -764,7 +1073,7 @@ def load_spatialite(conn):
     conn.create_function(
         "GFO_ReducePrecision",
         -1,
-        sqlite_userdefined.gfo_reduceprecision,
+        _sqlite_userdefined.gfo_reduceprecision,
         deterministic=True,
     )
 
