@@ -13,6 +13,31 @@ class ConfigOptions:
     """
 
     @classproperty
+    def copy_layer_sqlite_direct(cls) -> bool:
+        """Should copy_layer use sqlite directly when possible.
+
+        This is significantly faster than using GDAL for large datasets. At the moment
+        only used for .gpkg files.
+
+        Returns:
+            bool: True to use sqlite directly to copy layers. Defaults to True.
+        """
+        return get_bool("GFO_COPY_LAYER_SQLITE_DIRECT", default=True)
+
+    @classproperty
+    def io_engine(cls):
+        """The IO engine to use."""
+        io_engine = os.environ.get("GFO_IO_ENGINE", default="pyogrio").strip().lower()
+        supported_values = ["pyogrio", "fiona"]
+        if io_engine not in supported_values:
+            raise ValueError(
+                f"invalid value for configoption <GFO_IO_ENGINE>: '{io_engine}', "
+                f"should be one of {supported_values}"
+            )
+
+        return io_engine
+
+    @classproperty
     def on_data_error(cls) -> str:
         """The preferred action when a data error occurs.
 
@@ -35,16 +60,11 @@ class ConfigOptions:
         supported_values = ["raise", "warn"]
         if value_cleaned not in supported_values:
             raise ValueError(
-                f"invalid value for configoption <GFO_ON_DATA_ERROR>: {value}, should "
-                f"be one of {supported_values}"
+                f"invalid value for configoption <GFO_ON_DATA_ERROR>: '{value}', "
+                f"should be one of {supported_values}"
             )
 
         return value_cleaned
-
-    @classproperty
-    def io_engine(cls):
-        """The IO engine to use."""
-        return os.environ.get("GFO_IO_ENGINE", default="pyogrio").strip().lower()
 
     @classproperty
     def remove_temp_files(cls) -> bool:
@@ -60,14 +80,22 @@ class ConfigOptions:
         """The type of workers to use for parallel processing.
 
         Supported values (case insensitive):
-            - "thread": use threads when processing in parallel.
-            - "process": use processes when processing in parallel.
+            - "threads": use threads when processing in parallel.
+            - "processes": use processes when processing in parallel.
             - "auto": determine the type automatically.
 
         Returns:
             str: the type of workers to use. Defaults to "auto".
         """
-        return os.environ.get("GFO_WORKER_TYPE", default="auto").strip().lower()
+        worker_type = os.environ.get("GFO_WORKER_TYPE", default="auto").strip().lower()
+        supported_values = ["threads", "processes", "auto"]
+        if worker_type not in supported_values:
+            raise ValueError(
+                f"invalid value for configoption <GFO_WORKER_TYPE>: '{worker_type}', "
+                f"should be one of {supported_values}"
+            )
+
+        return worker_type
 
 
 def get_bool(key: str, default: bool) -> bool:
