@@ -1,20 +1,15 @@
-import os
+"""General helper functions, specific for geofileops."""
 
 from geofileops.helpers._configoptions_helper import ConfigOptions
 
 
-def use_threads(input_layer_featurecount: int) -> bool:
+def worker_type_to_use(input_layer_featurecount: int) -> str:
     worker_type = ConfigOptions.worker_type
-    if worker_type == "thread":
-        return True
-    elif worker_type == "process":
-        return False
+    if worker_type in ("threads", "processes"):
+        return worker_type
 
-    if os.name != "nt":
-        return False
-
-    # Processing in threads is 2x faster for small datasets on Windows
+    # Processing in threads is 2x faster for small datasets
     if input_layer_featurecount <= 100:
-        return True
+        return "threads"
 
-    return False
+    return "processes"
